Move producer/files.rs and producer/jobs.rs to transfer/*
