+++ peernode/src/peer-node/client.rs
@@ -3,6 +3,7 @@ pub mod consumer;
 pub mod grpc;
 pub mod producer;
 pub mod store;
+mod transfer;
 
 use std::io::{self, Write};
 use store::Configurations;
